# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 연금술사 Prototype API
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/artifacts:
    get:
      summary: Artifact List
      operationId: artifact_list_api_v1_projects__project_id__artifacts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artifacts/{artifact_id}:
    get:
      summary: Detail
      operationId: detail_api_v1_artifacts__artifact_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artifacts/{artifact_id}/download:
    get:
      summary: Download
      operationId: download_api_v1_artifacts__artifact_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      - name: version_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artifacts/{artifact_id}/reviews:
    post:
      summary: Review Artifact
      operationId: review_artifact_api_v1_artifacts__artifact_id__reviews_post
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artifacts/{artifact_id}/approve:
    post:
      summary: Approve Artifact
      operationId: approve_artifact_api_v1_artifacts__artifact_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/artifacts/{artifact_id}/regenerate:
    post:
      summary: Regenerate
      operationId: regenerate_api_v1_artifacts__artifact_id__regenerate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Artifact Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateInput'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RegenerateInput:
      properties:
        slide_id:
          type: string
          format: uuid
          title: Slide Id
      additionalProperties: false
      type: object
      required:
      - slide_id
      title: RegenerateInput
...
    ReviewInput:
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
          - request-change
          title: Decision
        comment:
          type: string
          maxLength: 2000
          title: Comment
          default: ''
        version:
          type: integer
          minimum: 1.0
          title: Version
      additionalProperties: false
      type: object
      required:
      - decision
      - version
      title: ReviewInput
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer